+++ scripts/egp32/executeProposal.ts
@@ -18,7 +18,7 @@ export async function executeProposal(signer: Wallet | undefined) {
   const { coreVoting } = addressesJson.addresses;
   const coreVotingContract = CoreVoting__factory.connect(coreVoting, signer);
 
-  const rawdata = fs.readFileSync("scripts/egp31/proposalInfo.json");
+  const rawdata = fs.readFileSync("scripts/egp32/proposalInfo.json");
   const proposalInfo: ProposalInfo = JSON.parse(rawdata.toString());
   const { proposalId, targets, callDatas } = proposalInfo;
