+++ scripts/egp32/proposalInfo.json
@@ -1,5 +1,5 @@
 {
-  "proposalId": 18,
+  "proposalId": 19,
   "votingVaults": [
     "0x02Bd4A3b1b95b01F2Aa61655415A5d3EAAcaafdD"
   ],
@@ -20,6 +20,6 @@
     "0x8522d1b20000000000000000000000005c6d51ecba4d8e4f20373e3ce96a62342b125d6d000000000000000000000000000000000000000000034f086f3b33b684000000000000000000000000000000f6094c3a380ad6161fb8240f3043392a0e427cac"
   ],
   "callHashTimelock": "0x7dac2ee7584cafc6693569b85f2af32a0702d5501c8f1f37e43ea0ee003b5cc7",
-  "lastCall": 20272718,
+  "lastCall": 20521518,
   "ballot": "0"
 }
